most of backend support for LMS mode; move configs into a defaults.py

+++ amplipi/streams.py
@@ -1124,9 +1124,10 @@ class FMRadio(BaseStream):
 
 class LMS(PersistentStream):
   """ An LMS Stream using squeezelite"""
-  def __init__(self, name: str, server: Optional[str] = None, disabled: bool = False, mock: bool = False):
+  def __init__(self, name: str, server: Optional[str] = None, disabled: bool = False, mock: bool = False, deallocatable: bool = True):
     super().__init__('lms', name, disabled=disabled, mock=mock)
     self.server : Optional[str] = server
+    self.deallocatable : bool = deallocatable
 
   def is_persistent(self):
     return True
@@ -1186,9 +1187,10 @@ class LMS(PersistentStream):
       print(f'error starting lms: {exc}')
 
   def _deactivate(self):
-    if self._is_running():
-      self.proc.kill()
-    self.proc = None
+    if self.deallocatable:
+      if self._is_running():
+        self.proc.kill()
+      self.proc = None
 
   def info(self) -> models.SourceInfo:
     source = models.SourceInfo(
@@ -1198,6 +1200,10 @@ class LMS(PersistentStream):
       track='check LMS for song info',
     )
     return source
+
+  def disconnect(self):
+    if self.deallocatable:
+      super().disconnect()
 
 class Bluetooth(BaseStream):
   """ A source for Bluetooth streams, which requires an external Bluetooth USB dongle """
@@ -1334,7 +1340,7 @@ def build_stream(stream: models.Stream, mock=False) -> AnyStream:
   if stream.type == 'fmradio':
     return FMRadio(name, args['freq'], args.get('logo'), disabled=disabled, mock=mock)
   if stream.type == 'lms':
-    return LMS(name, args.get('server'), disabled=disabled, mock=mock)
+    return LMS(name, args.get('server'), deallocatable=args.get('deallocatable', True), disabled=disabled, mock=mock)
   elif stream.type == 'bluetooth':
     return Bluetooth(name, disabled=disabled, mock=mock)
   raise NotImplementedError(stream.type)
